swap comments to use utteranc.es, remove outdated components
