+++ src/routes/(main)/[slug]/+page.svelte
@@ -77,23 +77,9 @@ $: canonical = SITE_URL + $page.url.pathname;
 </article>
 
 <div class="mx-auto mb-16 flex w-full max-w-2xl flex-col items-start justify-center px-4 sm:px-8">
-	<!-- <div class="mx-auto mb-16 w-fit max-w-2xl flex flex-col justify-center px-4 sm:px-8"> -->
-	<div
-		class="mb-12 w-fit self-center border-t-2 border-b-2 border-red-600 p-4 p-4 text-gray-700 dark:border-blue-300 dark:text-gray-400"
-	>
-		{#if json.ghMetadata.reactions.total_count > 0}
-			Reactions: <Reactions
-				issueUrl={json.ghMetadata.issueUrl}
-				reactions={json.ghMetadata.reactions}
-			/>
-		{:else}
-			<a class="gh-link" href={json.ghMetadata.issueUrl}>Leave a reaction </a>
-			if you liked this post! 🧡
-		{/if}
-	</div>
-	<div class="mb-8 w-full">
-		<Comments ghMetadata={json.ghMetadata} />
-	</div>
+<hr class="mt-2 w-full border-t-2 border-red-600 dark:border-blue-300" />
 
-	<!-- <Newsletter /> -->
+	<div class="mb-8 w-full">
+		<Comments issueNumber={json.issueNumber} />
+	</div>
 </div>
